refactor: 게임 실행 로직 컨트롤러에서 관리
